Feat: Allow cloudinary folders to be prefixed with item key.
The idea here is that cloudinary images or cloudinary image collections
which occur on a model with a unique key, we can add that unique key to
the cloudinary folder prefix.

ex. You have a model for post, and that models key is taken from the
name of the post. You could set prefixWithKey: true on your cloudinary
field and the generated folder prefix would be /<field.list.path>/<key>/<field.path>
instead of the usual /<field.list.path>/<field.path>

This is definitely in need of improvement such that any field can
be defined, but as of know this gets jg-keystone working for its
intended project.